M2063: signals were accepted and never delivered -- four fixes, four reverts
`rt_sigaction` and `rt_sigprocmask` were `r->rax = 0; /* accepted-and-
ignored for now */`, and kill/tkill/tgkill said, in the source, "other
signals: accepted, undelivered". That is the twelfth and thirteenth
"granted in name only" defect in this campaign, and together they made
every other piece of signal support pointless -- nothing could raise
anything, so none of it could ever run.

The worst-behaved part was the QUERY. `rt_sigaction(sig, NULL, &old)`
asks what is installed, and answering 0 without writing `old` leaves the
caller reading its own uninitialised stack as a function pointer; a
runtime that saves a handler to restore it later then installs garbage.
SIG_IGN was not recorded either, so a program that asked for a signal to
be ignored kept being interrupted by it -- and `sig_handler[sig] = 1`
would have been a jump to virtual address 1.

FOUR INDEPENDENT FIXES. Each was verified by reverting it alone:

1. rt_sigaction/rt_sigprocmask/rt_sigpending/rt_sigreturn implemented
   against the real machinery -- which already existed: app_sigaction, a
   pending bitset, sigaltstack, an RT sigqueue. SIG_IGN is stored as
   Linux's own sentinel (1) and checked everywhere a handler is about to
   run; a SIG_IGN signal is DISCARDED rather than left pending for ever.
   `rt_sigreturn` matters on its own: glibc's `__restore_rt` is
   `syscall(15)`, so a handler could have been entered and never
   returned. Reverted: 6 of 9 checks fail.

2. kill/tkill/tgkill route through a new `app_raise_signal_to`, which
   decides in app.c where the knowledge lives -- handler, SIG_IGN, or
   the default action -- rather than in the ABI layer. Reverted: 3 fail.

3. `app_deliver_pending` on the LINUX syscall return. The native return
   (syscall.c) and the interrupt return (interrupts.c) both did this;
   the Linux entry never did, so a process could install a handler,
   raise the signal, and return from the raising syscall straight past
   its own handler. Reverted: 2 fail.

4. THE SIGSET BIT NUMBERING. Linux's `sigmask(sig)` is `1 << (sig - 1)`
   and glibc's `sigaddset` agrees, so SIGWINCH (28) is bit 27 in every
   sigset crossing the boundary; our masks are indexed by the signal
   number itself. One shift either way, and getting it wrong is silent
   and total: a program blocks SIGWINCH, we record SIGPROF as blocked,
   and the signal it asked us to hold back is delivered immediately.
   Same class as M1967's struct statx offsets -- only an exact-value
   assertion finds it. Reverted: 2 fail.

AND THE NUMBERS NOW MATCH LINUX. SIGWINCH was 24 where Linux says 28,
and SIGRTMIN was 28, so a Linux signal number could not be passed
through: a program asking for SIGXCPU (24) would have been handed
SIGWINCH. SIGWINCH -> 28, SIGRTMIN -> 34, SIGXCPU/SIGXFSZ/SIGVTALRM/
SIGPROF at Linux's numbers, `APP_NSIG` 32 -> 65 and every mask widened
to 64 bits (Linux has 64 signals, and glibc uses SIGRTMIN+0..2 for
setxid and thread cancellation, which a 32-signal table cannot hold).

That renumbering found a fifth bug on its own: app.c carried LOCAL
`#define SIGXCPU 25` / `SIGXFSZ 26` overrides, added when 24 was
SIGWINCH. They silently shadowed the header, so the RLIMIT_CPU timer
raised the wrong signal -- and with signals real for the first time,
that hung the whole ABI suite at the subprocess probe. A redefinition
warning nobody was reading.

+++ kernel/syscall.c
@@ -2247,7 +2247,7 @@ void syscall_dispatch(struct registers *r) {
         int nfds = (int)r->rsi; long timeout = (long)r->rdx;
         if (nfds < 1 || nfds > 64 || !ubuf(r->rdi, (uint64_t)(unsigned)nfds * sizeof(struct pollfd))) { r->rax = (uint64_t)-1; break; }
         app_t *self = app_current();
-        uint32_t old_mask = app_sigprocmask(2 /* SIG_SETMASK */, (uint32_t)r->r10);
+        uint64_t old_mask = app_sigprocmask(2 /* SIG_SETMASK */, (uint64_t)r->r10);
         app_kill_check();
         __asm__ volatile("sti");
         uint64_t start = timer_ms();
@@ -2665,7 +2665,7 @@ void syscall_dispatch(struct registers *r) {
         struct epoll_event *out = (struct epoll_event *)r->rsi;
         int maxev = (int)r->rdx; long timeout = (long)r->r10;
         if (maxev < 1 || maxev > 64 || !ubuf(r->rsi, (uint64_t)(unsigned)maxev * sizeof(struct epoll_event))) { r->rax = (uint64_t)-1; break; }
-        uint32_t old_mask = app_sigprocmask(2 /* SIG_SETMASK */, (uint32_t)r->r8);
+        uint64_t old_mask = app_sigprocmask(2 /* SIG_SETMASK */, (uint64_t)r->r8);
         app_kill_check();
         __asm__ volatile("sti");
         uint64_t start = timer_ms(); int k = 0; int interrupted = 0;
@@ -2692,7 +2692,7 @@ void syscall_dispatch(struct registers *r) {
         r->rax = (uint64_t)app_getdents64((void *)r->rdi, (unsigned long)r->rsi, (int)r->rdx);
         break;
     case SYS_sigprocmask:                  /* (how, set): block/unblock signals; returns the old mask (M1208) */
-        r->rax = (uint64_t)app_sigprocmask((int)r->rdi, (uint32_t)r->rsi);
+        r->rax = (uint64_t)app_sigprocmask((int)r->rdi, (uint64_t)r->rsi);
         break;
     case SYS_sigpending:                   /* (): the pending (raised-but-blocked) signal set (M1209) */
         r->rax = (uint64_t)app_sigpending();
